Do not crash on workspace/didChangeConfiguration
The eslint-server LSP server sends the following message to the client:

    {
        "jsonrpc": "2.0",
        "params": {
            "registrations": [
                {
                    "registerOptions": {},
                    "method": "workspace/didChangeConfiguration",
                    "id": "4a56f5ca-7188-4f4c-a366-652d6f9d63aa"
                }
            ]
        },
        "method": "client/registerCapability",
        "id": 0
    }

This crashes lsp-test because it asserts that registerOptions is null
(but an empty object is not null).

Make lsp-test accept either null or an object.

Question to reviewers: What's the right way to handle this? Should we
make a type similar to WorkspaceSymbolRegistrationOptions? Should we
change Empty to accept empty objects in addition to null?

Question to reviewers: Where do we write tests for this change?